+++ easyarmorstands-display/src/main/java/me/m56738/easyarmorstands/display/element/DisplayElement.java
@@ -6,14 +6,17 @@ import me.m56738.easyarmorstands.api.editor.button.Button;
 import me.m56738.easyarmorstands.api.editor.node.Node;
 import me.m56738.easyarmorstands.api.property.PropertyContainer;
 import me.m56738.easyarmorstands.api.util.BoundingBox;
+import me.m56738.easyarmorstands.display.editor.BlockDisplayOffsetProvider;
 import me.m56738.easyarmorstands.display.editor.DisplayOffsetProvider;
 import me.m56738.easyarmorstands.display.editor.DisplayRotationProvider;
 import me.m56738.easyarmorstands.display.editor.node.DisplayRootNode;
 import me.m56738.easyarmorstands.editor.EntityPositionProvider;
+import me.m56738.easyarmorstands.editor.OffsetProvider;
 import me.m56738.easyarmorstands.element.SimpleEntityElement;
 import me.m56738.easyarmorstands.element.SimpleEntityElementType;
 import me.m56738.easyarmorstands.lib.joml.Vector3d;
 import me.m56738.easyarmorstands.util.Util;
+import org.bukkit.entity.BlockDisplay;
 import org.bukkit.entity.Display;
 import org.jetbrains.annotations.NotNull;
 
@@ -29,7 +32,7 @@ public class DisplayElement<T extends Display> extends SimpleEntityElement<T> {
     public @NotNull Button createButton(@NotNull Session session) {
         PropertyContainer properties = getProperties();
         return new BoundingBoxButton(session, this,
-                new EntityPositionProvider(properties, new DisplayOffsetProvider(properties)),
+                new EntityPositionProvider(properties, getOffsetProvider(properties)),
                 new DisplayRotationProvider(properties));
     }
 
@@ -40,7 +43,15 @@ public class DisplayElement<T extends Display> extends SimpleEntityElement<T> {
 
     @Override
     public @NotNull DisplayToolProvider getTools(@NotNull PropertyContainer properties) {
-        return new DisplayToolProvider(properties);
+        return new DisplayToolProvider(properties, getOffsetProvider(properties));
+    }
+
+    private OffsetProvider getOffsetProvider(PropertyContainer properties) {
+        if (entity instanceof BlockDisplay) {
+            return new BlockDisplayOffsetProvider(properties);
+        } else {
+            return new DisplayOffsetProvider(properties);
+        }
     }
 
     @Override