+++ easyarmorstands-fancyholograms/src/main/java/me/m56738/easyarmorstands/fancyholograms/element/HologramToolProvider.java
@@ -7,13 +7,16 @@ import me.m56738.easyarmorstands.api.editor.tool.ToolContext;
 import me.m56738.easyarmorstands.api.property.PropertyContainer;
 import me.m56738.easyarmorstands.display.editor.tool.DisplayAxisScaleTool;
 import me.m56738.easyarmorstands.display.editor.tool.DisplayScaleTool;
+import me.m56738.easyarmorstands.editor.EntityPositionProvider;
+import me.m56738.easyarmorstands.editor.OffsetProvider;
 import me.m56738.easyarmorstands.element.SimpleEntityToolProvider;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class HologramToolProvider extends SimpleEntityToolProvider {
-    public HologramToolProvider(PropertyContainer properties) {
+    public HologramToolProvider(PropertyContainer properties, OffsetProvider offsetProvider) {
         super(properties);
+        positionProvider = new EntityPositionProvider(properties, offsetProvider);
     }
 
     @Override
@@ -23,7 +26,7 @@ public class HologramToolProvider extends SimpleEntityToolProvider {
 
     @Override
     public @Nullable AxisScaleTool scale(@NotNull ToolContext context, @NotNull Axis axis) {
-        if (context.position() == position() && context.rotation() == rotation()) {
+        if (context.rotation() == rotation()) {
             return new DisplayAxisScaleTool(context, properties, axis);
         }
         return super.scale(context, axis);