+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Changed control to switch modes to Q (dropping item).
   * Previously Shift + Right-click.
 * `/eas give` selects the tool if you already have it in your inventory.
+* Move block display pivot to the center
 
 # Fixes